adjust test order aviod segmentfault
