+++ test/ir/pir/test_standalone_pir.py
@@ -12,8 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import os
-import tempfile
 import unittest
 
 import numpy as np
@@ -285,47 +283,6 @@ class TestPirPrint(unittest.TestCase):
         gold_res = np.ones([2, 2], dtype="float32") * 2
 
         np.testing.assert_array_equal(out[0], gold_res)
-
-
-class TestJitSaveOp(unittest.TestCase):
-    def setUp(self):
-        self.temp_dir = tempfile.TemporaryDirectory()
-        self.model_path = os.path.join(self.temp_dir.name, "pir_save_load")
-
-    def tearDown(self):
-        self.temp_dir.cleanup()
-
-    def test_with_pir(self):
-        with paddle.pir_utils.OldIrGuard():
-            paddle.disable_static()
-            linear = paddle.nn.Linear(10, 10)
-            path = os.path.join(self.model_path, "linear")
-
-            paddle.jit.save(
-                linear,
-                path,
-                input_spec=[paddle.static.InputSpec([10, 10], 'float32', 'x')],
-            )
-
-            paddle.enable_static()
-            place = (
-                paddle.CUDAPlace(0)
-                if paddle.is_compiled_with_cuda()
-                else paddle.CPUPlace()
-            )
-
-            exe = paddle.static.Executor(place)
-
-            [
-                inference_program,
-                feed_target_names,
-                fetch_targets,
-            ] = paddle.static.io.load_inference_model(
-                self.model_path,
-                executor=exe,
-                model_filename="linear.pdmodel",
-                params_filename="linear.pdiparams",
-            )
 
 
 class TestPirConcatDygraph(unittest.TestCase):
